+++ ui/frontend/src/main/src/app/inventur/inventur.service.ts
@@ -4,8 +4,8 @@ import { Observable } from 'rxjs/Observable';
 import { NgRedux, select } from '@angular-redux/store';
 import { CommandGatewayService } from '../shared/command-gateway/command-gateway.service';
 import { QueryGatewayService } from '../shared/query-gateway/query-gateway.service';
-import { QueryType } from '../shared/query-gateway/query-type';
-import { ResultType } from '../shared/query-gateway/result-type';
+import { QueryType } from './query-type';
+import { ResultType } from './result-type';
 import { Inventar } from './inventar';
 import { AppState, InventurState } from '../store/model';
 import { CommandType } from './command-type';
