+++ ui/frontend/src/main/src/app/inventur/result-type.ts
