+++ ui/frontend/src/main/src/app/store/inventur-reducers.ts
@@ -2,7 +2,7 @@ import { QueryGatewayActionType } from '../shared/query-gateway/query-reducers';
 import { INVENTUR_INITIAL_STATE, InventurState } from './model';
 import { CommandGatewayActionType } from '../shared/command-gateway/command-gateway-actions.service';
 import { CommandType } from '../inventur/command-type';
-import { QueryType } from '../shared/query-gateway/query-type';
+import { QueryType } from '../inventur/query-type';
 
 export function inventurReducer(state: InventurState = INVENTUR_INITIAL_STATE, action): InventurState {
     switch (action.type) {
